+++ CHANGELOG.md
@@ -3,5 +3,17 @@
 # voml-intellij Changelog
 
 ## [Unreleased]
+
+## [v0.1.2]
+
+### Fix
+- The scope tag is not highlighted correctly
+
+## [v0.1.1]
+### Added
+- Syntax highlight for `@inherit` statement
+- Syntax highlight for `@include` statement
+
+## [v0.1.0]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)